inliear regression, scoring integer

+++ include/frame.h
@@ -38,7 +38,7 @@ std::tuple<std::vector<cv::KeyPoint>, std::vector<cv::KeyPoint>, Mat, Mat> condi
 bool test_collinear(std::vector<cv::KeyPoint>kpts1, float  eps=5.5f);
 cv::Point3f point2line(Point2f,Point2f);
 float dist(Point2f, Point3f);
-std::tuple<float, std::vector<int>> concensus(std::vector<cv::KeyPoint> kpts1, std::vector<cv::KeyPoint> kpts2, std::vector<cv::DMatch> matches, Mat H, Mat invH, float eps=0.5f);
+std::tuple<unsigned int, std::vector<int>> concensus(std::vector<cv::KeyPoint> kpts1, std::vector<cv::KeyPoint> kpts2, std::vector<cv::DMatch> matches, Mat H, Mat invH, float eps=0.5f);
 Point2f Transformation(Mat, Point2f);
 float pointNorm(Point2f);
 std::tuple<Mat, float> ransac(std::vector<cv::KeyPoint> kpts1, std::vector<cv::KeyPoint> kpts2, std::vector<cv::DMatch> matches, float min=0.5f, float p=0.99, float s=4,float eps=0.5f);
